Start of Typewriter Git
It has been some time since this project was started. It was worked on every now and then.

First the basics of everything had been created: Drawing letters at positions using SDL.
After that basic functionality has been added. Eventually things had been at the same level as the ingame typewriter.
As next step handling newline was partially implemented.

However a fundamental flaw in the typewriters base capability have been encountered. Where the cursor would be in the next row due to auto-incrementation, while it should be at the end of the current line.

This messes up a bunch of things. A major rewrite has to be done in code and hardware later.

The current newline implementation is incomplete and partially broken as not all edge cases have been handled yet. There also exist cavemen debugging statements.
But this is the second best chance to start a Git.