Add one more cursor column and fix everything related
Previously there only had been as many cursor columns as letters.
But a new line has the cursor before the first letter slot and a full line after the last letter. A wrapped line never has the cursor in the first column.
This old way was taken directly from the previous LW hardware typewriter.

Besides that, this commit fixes everything related to newlines. Which was not possible previously without that additional cursor position.

The text editor should be fully functional right now - with all the original features.